Add a cIExemplarPatchingServer GZCOM interface
This interface allows DLLs that dynamically load and unload DBPF files
to request a new scan for exemplar patches.
For example, the Extended Terrain DLL could trigger a new scan for
exemplar patches after it loads a new terrain mod.

The ExemplarPatcher class now uses a mutex to prevent the exemplar
patch list from being read and modified at the same time.
This is necessary because there is a rare chance that a DLL could request
a new scan for exemplar patches at the same time the game's main thread
is applying patches.

Because SC4 is a mostly single threaded game and exemplar patch reloads
should be rare, there likely wouldn't be any major benefit to changing
the exemplar patch list over to boost::concurrent_flat_map.