Fix an ExemplarPatcher synchronization object crash
A user was experiencing a crash with the CRT std::mutex type, see https://community.simtropolis.com/forums/topic/763167-resource-loading-hooks-dll-for-simcity-4/?do=findComment&comment=1808293.
The crash was internal to the CRT, and I have not been able to
reproduce it.
So it may be something specific to their system.

The fix was replacing std::mutex with wil::critical_section.